Change migration to delete if exists

+++ db/migrate/20180223145638_drop_table_models_from_schema.rb
@@ -1,5 +1,5 @@
 class DropTableModelsFromSchema < ActiveRecord::Migration[5.1]
   def change
-  	drop_table :models
+  	drop_table :models if self.table_exists?("models")
   end
 end
